Refactor grid creation
The grid now builds itself in the top orientation. each inserted ship is then rotated based on that intial position. This reduced a lot of code and means the grid built, always has the first row closest to the player. This will make filling the grid with on-screen enemies much easier.